Introducing 'Old cURL' mode
Some environments may have outdated versions of some tools. Because
some behaviour in earlier versions of tools will not match expected
behaviour when implementing scripts again current versions, it may
become required to use a different mechanism to obtain the same data
format.

This PR introduces a bash-scripted urlencode equivalent for cases
where a machine's cURL is outdated and cannot produce an urlencoded
string the way codecov expects it to.